+++ src/app/components/availability/context/availability-context.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { createContext, Dispatch, SetStateAction, useContext } from "react";
-import { useAvailabilityState } from "@/app/components/availability/use-availability-state";
+import { useAvailabilityState } from "@/app/components/availability/context/use-availability-state";
 import {
     AvailabilityBlockType,
     SelectionStateType,
