+++ src/app/components/availability/context/use-availability-state.ts
